+++ lib/features/course_selection/data/dtos/course_dto.dart
@@ -38,7 +38,7 @@ class CourseResponseDto {
   final String serialNo;
   final String classNo;
   final String title;
-  final double credit;
+  final int credit;
   final String? passwordCard;
   @JsonKey(defaultValue: <String>[])
   final List<String> teachers;
