+++ test/pages/course_selection_page_test.dart
@@ -158,7 +158,7 @@ class FakeCourseRepository implements CourseRepository {
     String? departmentId,
     String? collegeId,
     String? courseType,
-    List<double>? credits,
+    List<int>? credits,
     bool? hasVacancy,
     List<String>? classTimes,
     int offset = 0,
@@ -193,7 +193,7 @@ class CourseSearchRequest {
 
   final String? keyword;
   final String? courseType;
-  final List<double>? credits;
+  final List<int>? credits;
   final bool? hasVacancy;
   final List<String>? classTimes;
   final int offset;
